producers/copernicus_insitu/requirements.txt
copernicusmarine
pandas
confluent-kafka
pyyaml
